Add option for enforcing usage fork()
There is a bug when running nodemon in a Docker container using spawn().

If there is an asychronous SIGUSR2 handler registered, nodemon will not
wait for the asynchronous operations before spawning a new process.

This new option is a way to force nodemon to use fork(), avoiding the
spawn() bug.